Bad Dude shoots at the player when in view
By 'in view', I mean when the player is in the same direction the Bad Dude is facing. This commit has a LOT of changes, because I realized I made an egregious error in the MBC code when defining GOTO, SUBR, and branch statements. There'll be a more informative PR available at some point, but the gist is that 6 bits can only go as high as 63, so if you tried to use any of these instructions to get to a line past 63, the results were unpredictable. I also added some helpful logging functions.